+++ src/lib/WP3StylesListener.cpp
@@ -36,8 +36,8 @@ WP3StylesListener::WP3StylesListener(std::list<WPXPageSpan> &pageList, WPXTableL
 	m_currentPage(WPXPageSpan()),
 	m_tableList(tableList), 
 	m_currentTable(0),
-	m_tempMarginLeft(1.0f),
-	m_tempMarginRight(1.0f),
+	m_tempMarginLeft(1.0),
+	m_tempMarginRight(1.0),
 	m_currentPageHasContent(false),
 	m_isSubDocument(false),
 	m_subDocuments(subDocuments),
@@ -82,7 +82,7 @@ void WP3StylesListener::insertBreak(const uint8_t breakType)
 				if (m_pageListHardPageMark == m_pageList.end())
 					m_pageListHardPageMark--;
 			}
-			m_currentPage = WPXPageSpan(m_pageList.back(), 0.0f, 0.0f);
+			m_currentPage = WPXPageSpan(m_pageList.back(), 0.0, 0.0);
 			m_currentPage.setPageSpan(1);
 			m_currentPageHasContent = false;
 			break;
@@ -111,7 +111,7 @@ void WP3StylesListener::pageMarginChange(const uint8_t side, const uint16_t marg
 {
 	if (!isUndoOn()) 
 	{
-		float marginInch = (float)((double)margin / (double)WPX_NUM_WPUS_PER_INCH);
+		double marginInch = (double)((double)margin / (double)WPX_NUM_WPUS_PER_INCH);
 		switch(side)
 		{
 			case WPX_TOP:
@@ -132,7 +132,7 @@ void WP3StylesListener::marginChange(const uint8_t side, const uint16_t margin)
 			return; // do not collect L/R margin information in sub documents
 
 		std::list<WPXPageSpan>::iterator Iter;
-		float marginInch = (float)((double)margin / (double)WPX_NUM_WPUS_PER_INCH);
+		double marginInch = (double)((double)margin / (double)WPX_NUM_WPUS_PER_INCH);
 		switch(side)
 		{
 			case WPX_LEFT:
@@ -172,8 +172,8 @@ void WP3StylesListener::pageFormChange(const uint16_t length, const uint16_t wid
 {
 	if (!isUndoOn())
 	{
-		float lengthInch = (float)((double)length / (double)WPX_NUM_WPUS_PER_INCH);
-		float widthInch = (float)((double)width / (double)WPX_NUM_WPUS_PER_INCH);
+		double lengthInch = (double)((double)length / (double)WPX_NUM_WPUS_PER_INCH);
+		double widthInch = (double)((double)width / (double)WPX_NUM_WPUS_PER_INCH);
 		if (!m_currentPageHasContent)
 		{
 			m_currentPage.setFormLength(lengthInch);
