+++ src/lib/WP5FontGroup.h
@@ -39,7 +39,7 @@ protected:
 private:
 	uint8_t m_red, m_green, m_blue;
 	uint8_t m_fontNumber;
-	float m_fontSize;
+	double m_fontSize;
 };
 
 #endif /* WP5FONTGROUP_H */
